[Add] #182 - SettingView 파일 생성
